+++ src/marqo_commons/model_registry/model_properties_object.py
@@ -7,7 +7,7 @@ class ModelProperties(BaseModel):
     dimensions: int = Field(..., title="Model dimensions")
     notes: str = Field(..., title="Model notes")
     type: str = Field(..., title="Model types")
-    memory_size: int = Field(..., title="Model memory size")
+    memory_size: float = Field(..., title="Model memory size")
     modality: list[Modality] = Field(..., title="Model modality")
     vector_numeric_type: VectorNumericType = Field(..., title="Model vector numeric type")
